+++ src/goose.rs
@@ -286,8 +286,11 @@
 //! limitations under the License.
 
 use downcast_rs::{impl_downcast, Downcast};
+use http::header::{Entry, IntoHeaderName};
 use http::method::Method;
-use reqwest::{header, Client, ClientBuilder, RequestBuilder, Response};
+use http::HeaderValue;
+use reqwest::cookie::{CookieStore, Jar};
+use reqwest::{cookie, header, Client, RequestBuilder, Response};
 use serde::{Deserialize, Serialize};
 use std::hash::{Hash, Hasher};
 use std::sync::Arc;
@@ -302,9 +305,6 @@ use crate::metrics::{
     GooseCoordinatedOmissionMitigation, GooseMetric, GooseRawRequest, GooseRequestMetric,
 };
 use crate::{GooseConfiguration, GooseError, WeightedGooseTasks};
-
-/// By default Goose sets the following User-Agent header when making requests.
-static APP_USER_AGENT: &str = concat!(env!("CARGO_PKG_NAME"), "/", env!("CARGO_PKG_VERSION"));
 
 /// `task!(foo)` expands to `GooseTask::new(foo)`, but also does some boxing to work around a limitation in the compiler.
 #[macro_export]
@@ -853,22 +853,21 @@ pub struct GooseUser {
     pub(crate) task_name: Option<String>,
 
     session_data: Option<Box<dyn GooseUserData>>,
+
+    cookie_store: Jar,
+
+    headers: header::HeaderMap,
 }
 impl GooseUser {
     /// Create a new user state.
     pub fn new(
+        client: Client,
         task_sets_index: usize,
         base_url: Url,
         configuration: &GooseConfiguration,
         load_test_hash: u64,
     ) -> Result<Self, GooseError> {
         trace!("new GooseUser");
-        let client = Client::builder()
-            .user_agent(APP_USER_AGENT)
-            .cookie_store(true)
-            // Enable gzip unless `--no-gzip` flag is enabled.
-            .gzip(!configuration.no_gzip)
-            .build()?;
 
         Ok(GooseUser {
             started: Instant::now(),
@@ -887,12 +886,18 @@ impl GooseUser {
             slept: 0,
             task_name: None,
             session_data: None,
+            cookie_store: cookie::Jar::default(),
+            headers: header::HeaderMap::default(),
         })
     }
 
     /// Create a new single-use user.
-    pub fn single(base_url: Url, configuration: &GooseConfiguration) -> Result<Self, GooseError> {
-        let mut single_user = GooseUser::new(0, base_url, configuration, 0)?;
+    pub fn single(
+        client: Client,
+        base_url: Url,
+        configuration: &GooseConfiguration,
+    ) -> Result<Self, GooseError> {
+        let mut single_user = GooseUser::new(client, 0, base_url, configuration, 0)?;
         // Only one user, so index is 0.
         single_user.weighted_users_index = 0;
         // Do not throttle [`test_start`](../struct.GooseAttack.html#method.test_start) (setup) and
@@ -1537,8 +1542,7 @@ impl GooseUser {
         };
 
         let started = Instant::now();
-        let request = request_builder.build()?;
-
+        let mut request = request_builder.build()?;
         // String version of request path.
         let path = match Url::parse(&request.url().to_string()) {
             Ok(u) => u.path().to_string(),
@@ -1581,6 +1585,12 @@ impl GooseUser {
             self.started.elapsed().as_millis(),
             self.weighted_users_index,
         );
+
+        // Set cookies
+        self.add_cookie_header(&mut request);
+
+        // Set default headers
+        self.add_default_headers(&mut request);
 
         // Make the actual request.
         let response = self.client.execute(request).await;
@@ -1640,6 +1650,26 @@ impl GooseUser {
         }
 
         Ok(GooseResponse::new(request_metric, response))
+    }
+
+    // Add cookies from the cookie store.
+    fn add_cookie_header(&self, request: &mut reqwest::Request) {
+        if request.headers().get(header::COOKIE).is_none() {
+            if let Some(header) = self.cookie_store.cookies(request.url()) {
+                request.headers_mut().insert(header::COOKIE, header);
+            }
+        }
+    }
+
+    // insert default headers in the request headers
+    // without overwriting already appended headers.
+    fn add_default_headers(&self, request: &mut reqwest::Request) {
+        let headers = request.headers_mut();
+        for (key, value) in &self.headers {
+            if let Entry::Vacant(entry) = headers.entry(key) {
+                entry.insert(value.clone());
+            }
+        }
     }
 
     /// Tracks the time it takes for the current GooseUser to loop through all GooseTasks
@@ -2041,152 +2071,55 @@ impl GooseUser {
         Ok(())
     }
 
-    /// Manually build a
-    /// [`reqwest::Client`](https://docs.rs/reqwest/*/reqwest/struct.Client.html).
-    ///
-    /// By default, Goose configures two options when building a
-    /// [`reqwest::Client`](https://docs.rs/reqwest/*/reqwest/struct.Client.html). The first
-    /// configures Goose to report itself as the
-    /// [`user_agent`](https://docs.rs/reqwest/*/reqwest/struct.ClientBuilder.html#method.user_agent)
-    /// requesting web pages (ie `goose/0.11.2`). The second option configures
-    /// [`reqwest`](https://docs.rs/reqwest/) to
-    /// [store cookies](https://docs.rs/reqwest/*/reqwest/struct.ClientBuilder.html#method.cookie_store),
-    /// which is generally necessary if you aim to simulate logged in users.
-    ///
-    /// # Default configuration:
-    ///
-    /// ```rust
-    /// use reqwest::Client;
-    ///
-    /// static APP_USER_AGENT: &str = concat!(env!("CARGO_PKG_NAME"), "/", env!("CARGO_PKG_VERSION"));
-    ///
-    /// let builder = Client::builder()
-    ///   .user_agent(APP_USER_AGENT)
-    ///   .cookie_store(true)
-    ///   .gzip(true);
-    /// ```
-    ///
-    /// Alternatively, you can use this function to manually build a
-    /// [`reqwest::Client`](https://docs.rs/reqwest/*/reqwest/struct.Client.html).
-    /// with custom configuration. Available options are found in the
-    /// [`reqwest::ClientBuilder`](https://docs.rs/reqwest/*/reqwest/struct.ClientBuilder.html)
-    /// documentation.
-    ///
-    /// When manually building a
-    /// [`reqwest::Client`](https://docs.rs/reqwest/*/reqwest/struct.Client.html),
-    /// there are a few things to be aware of:
-    ///  - Manually building a client in [`test_start`](../struct.GooseAttack.html#method.test_start)
-    ///    will only affect requests made during test setup;
-    ///  - Manually building a client in [`test_stop`](../struct.GooseAttack.html#method.test_stop)
-    ///    will only affect requests made during test teardown;
-    ///  - A manually built client is specific to a single Goose thread -- if you are
-    ///    generating a large load test with many users, each will need to manually build their
-    ///    own client (typically you'd do this in a Task that is registered with
-    ///   [`GooseTask::set_on_start()`] in each Task Set requiring a custom client;
-    ///  - Manually building a client will completely replace the automatically built client
-    ///    with a brand new one, so any configuration, cookies or headers set in the previously
-    ///    built client will be gone;
-    ///  - You must include all desired configuration, as you are completely replacing Goose
-    ///    defaults. For example, if you want Goose clients to store cookies, you will have to
-    ///    include
-    ///    [`.cookie_store(true)`](https://docs.rs/reqwest/*/reqwest/struct.ClientBuilder.html#method.cookie_store).
-    ///
-    /// In the following example, the Goose client is configured with a different user agent,
-    /// sets a default header on every request, stores cookies, and supports gzip compression.
+    /// Custom cookies can be manually set to each [`GooseUser`](./goose/struct.GooseUser.html).
+    /// Those cookies will automatically been added to each request.
     ///
     /// ## Example
     /// ```rust
     /// use goose::prelude::*;
     ///
-    /// task!(setup_custom_client).set_on_start();
+    /// task!(custom_cookie).set_on_start();
     ///
-    /// async fn setup_custom_client(user: &mut GooseUser) -> GooseTaskResult {
-    ///     use reqwest::{Client, header};
-    ///
-    ///     // Build a custom HeaderMap to include with all requests made by this client.
-    ///     let mut headers = header::HeaderMap::new();
-    ///     headers.insert("X-Custom-Header", header::HeaderValue::from_str("custom value").unwrap());
-    ///
-    ///     // Build a custom client.
-    ///     let builder = Client::builder()
-    ///         .default_headers(headers)
-    ///         .user_agent("custom user agent")
-    ///         .cookie_store(true)
-    ///         .gzip(true);
-    ///
-    ///     // Assign the custom client to this GooseUser.
-    ///     user.set_client_builder(builder).await?;
-    ///
-    ///     Ok(())
-    /// }
-    /// ```
-    ///
-    /// # Alternative Compression Algorithms
-    /// Reqwest also supports
-    /// [`brotli`](https://docs.rs/reqwest/*/reqwest/struct.ClientBuilder.html#method.brotli) and
-    /// [`deflate`](https://docs.rs/reqwest/*/reqwest/struct.ClientBuilder.html#method.deflate) compression.
-    ///
-    /// To enable either, you must enable the features in your load test's `Cargo.toml`, for example:
-    /// ```text
-    /// reqwest = { version = "^0.11.4",  default-features = false, features = [
-    ///     "brotli",
-    ///     "cookies",
-    ///     "deflate",
-    ///     "gzip",
-    ///     "json",
-    /// ] }
-    /// ```
-    ///
-    /// Once enabled, you can add `.brotli(true)` and/or `.deflate(true)` to your custom
-    /// [`reqwest::Client::builder()`], following the documentation above.
-    ///
-    /// # Custom Cookies
-    /// Custom cookies can also be manually set when building a custom [`reqwest::Client`]. This requires
-    /// loading the [`GooseUser::base_url`] being load tested in order to properly build the cookie. Then
-    /// a custom [`reqwest::cookie::Jar`] is created and the custom cookie is added with
-    /// [`reqwest::cookie::Jar::add_cookie_str`]. Finally, the new cookie jar must be specified as the
-    /// [`reqwest::ClientBuilder::cookie_provider`] for the custom client.
-    ///
-    /// ## Example
-    /// ```rust
-    /// use reqwest::{cookie::Jar, Client};
-    /// use std::sync::Arc;
-    ///
-    /// use goose::prelude::*;
-    ///
-    /// task!(custom_cookie_with_custom_client).set_on_start();
-    ///
-    /// async fn custom_cookie_with_custom_client(user: &mut GooseUser) -> GooseTaskResult {
+    /// async fn custom_cookie(user: &mut GooseUser) -> GooseTaskResult {
     ///     // Prepare the contents of a custom cookie.
     ///     let cookie = "my-custom-cookie=custom-value";
     ///
-    ///     // Pre-load one or more cookies into a custom cookie jar to use with this client.
-    ///     let jar = Jar::default();
-    ///     jar.add_cookie_str(
+    ///     // Pre-load one or more cookies into the GooseUser.
+    ///     user.add_cookie_str(
     ///         cookie,
-    ///         &user.base_url,
     ///     );
-    ///
-    ///     // Build a custom client.
-    ///     let builder = Client::builder()
-    ///         .user_agent("example-loadtest")
-    ///         .cookie_store(true)
-    ///         .cookie_provider(Arc::new(jar))
-    ///         .gzip(true);
-    ///
-    ///     // Assign the custom client to this GooseUser.
-    ///     user.set_client_builder(builder).await?;
     ///
     ///     Ok(())
     /// }
     /// ```
-    pub async fn set_client_builder(
-        &mut self,
-        builder: ClientBuilder,
-    ) -> Result<(), GooseTaskError> {
-        self.client = builder.build()?;
+    pub fn add_cookie_str(&mut self, cookie: &str) {
+        self.cookie_store.add_cookie_str(cookie, &self.base_url);
+    }
 
-        Ok(())
+    /// Default headers can be manually set to each [`GooseUser`](./goose/struct.GooseUser.html).
+    /// The headers will automatically been added to each request made by the [`GooseUser`](./goose/struct.GooseUser.html).
+    ///
+    /// Default headers doesn't overwrite already appended headers.
+    /// ## Example
+    /// ```rust
+    /// use goose::prelude::*;
+    /// use reqwest::header;
+    ///
+    /// task!(add_default_header).set_on_start();
+    ///
+    /// async fn add_default_header(user: &mut GooseUser) -> GooseTaskResult {
+    ///
+    ///     // Build a custom HeaderMap to include with all requests made by this client.
+    ///     user.add_default_header("X-Custom-Header", header::HeaderValue::from_str("custom value").unwrap());
+    ///
+    ///     Ok(())
+    /// }
+    /// ```
+    pub fn add_default_header<K>(&mut self, key: K, val: HeaderValue)
+    where
+        K: IntoHeaderName,
+    {
+        self.headers.insert(key, val);
     }
 
     /// Some websites use multiple domains to serve traffic, redirecting depending on
@@ -2606,10 +2539,11 @@ mod tests {
     const EMPTY_ARGS: Vec<&str> = vec![];
 
     fn setup_user(server: &MockServer) -> Result<GooseUser, GooseError> {
+        let client = Client::builder().build().unwrap();
         let mut configuration = GooseConfiguration::parse_args_default(&EMPTY_ARGS).unwrap();
         configuration.co_mitigation = Some(GooseCoordinatedOmissionMitigation::Average);
         let base_url = get_base_url(Some(server.url("/")), None, None).unwrap();
-        GooseUser::single(base_url, &configuration)
+        GooseUser::single(client, base_url, &configuration)
     }
 
     #[test]
@@ -2797,9 +2731,10 @@ mod tests {
     #[tokio::test]
     async fn goose_user() {
         const HOST: &str = "http://example.com/";
+        let client = Client::builder().build().unwrap();
         let configuration = GooseConfiguration::parse_args_default(&EMPTY_ARGS).unwrap();
         let base_url = get_base_url(Some(HOST.to_string()), None, None).unwrap();
-        let user = GooseUser::new(0, base_url, &configuration, 0).unwrap();
+        let user = GooseUser::new(client.clone(), 0, base_url, &configuration, 0).unwrap();
         assert_eq!(user.task_sets_index, 0);
         assert_eq!(user.weighted_users_index, usize::max_value());
 
@@ -2826,7 +2761,7 @@ mod tests {
             Some("http://www.example.com/".to_string()),
         )
         .unwrap();
-        let user2 = GooseUser::new(0, base_url, &configuration, 0).unwrap();
+        let user2 = GooseUser::new(client.clone(), 0, base_url, &configuration, 0).unwrap();
 
         // Confirm the URLs are correctly built using the task_set_host.
         let url = user2.build_url("/foo").unwrap();
@@ -2885,7 +2820,7 @@ mod tests {
         // Confirm Goose can build a base_url that includes a path.
         const HOST_WITH_PATH: &str = "http://example.com/with/path/";
         let base_url = get_base_url(Some(HOST_WITH_PATH.to_string()), None, None).unwrap();
-        let user = GooseUser::new(0, base_url, &configuration, 0).unwrap();
+        let user = GooseUser::new(client, 0, base_url, &configuration, 0).unwrap();
 
         // Confirm the URLs are correctly built using the default_host that includes a path.
         let url = user.build_url("foo").unwrap();
@@ -2987,9 +2922,14 @@ mod tests {
             data: "foo".to_owned(),
         };
 
+        let client = Client::builder().build().unwrap();
         let configuration = GooseConfiguration::parse_args_default(&EMPTY_ARGS).unwrap();
-        let mut user =
-            GooseUser::single("http://localhost:8080".parse().unwrap(), &configuration).unwrap();
+        let mut user = GooseUser::single(
+            client,
+            "http://localhost:8080".parse().unwrap(),
+            &configuration,
+        )
+        .unwrap();
 
         user.set_session_data(session_data.clone());
 
@@ -3011,10 +2951,14 @@ mod tests {
         let session_data = CustomSessionData {
             data: "foo".to_owned(),
         };
-
+        let client = Client::builder().build().unwrap();
         let configuration = GooseConfiguration::parse_args_default(&EMPTY_ARGS).unwrap();
-        let mut user =
-            GooseUser::single("http://localhost:8080".parse().unwrap(), &configuration).unwrap();
+        let mut user = GooseUser::single(
+            client,
+            "http://localhost:8080".parse().unwrap(),
+            &configuration,
+        )
+        .unwrap();
 
         user.set_session_data(session_data);
 
@@ -3042,9 +2986,14 @@ mod tests {
             data: "foo".to_owned(),
         };
 
+        let client = Client::builder().build().unwrap();
         let configuration = GooseConfiguration::parse_args_default(&EMPTY_ARGS).unwrap();
-        let mut user =
-            GooseUser::single("http://localhost:8080".parse().unwrap(), &configuration).unwrap();
+        let mut user = GooseUser::single(
+            client,
+            "http://localhost:8080".parse().unwrap(),
+            &configuration,
+        )
+        .unwrap();
 
         user.set_session_data(session_data.clone());